refactor(peapods): restructure code base
