+++ jina/peapods/gateway/rest.py
@@ -4,7 +4,7 @@ from typing import Any
 from google.protobuf.json_format import MessageToDict
 
 from .grpc import GatewayPea
-from .servicer import GRPCServicer
+from .grpc.async_servicer import GRPCServicer
 from ...enums import RequestType
 from ...helper import configure_event_loop
 from ...importer import ImportExtensions
@@ -17,6 +17,7 @@ class RESTGatewayPea(GatewayPea):
     Unlike :class:`GatewayPea`, it does not support bi-directional streaming. Therefore, it is
     synchronous from the client perspective.
     """
+
     def loop_body(self):
         configure_event_loop()
         self.gateway = RESTGateway(host=self.args.host,
@@ -32,7 +33,6 @@ class RESTGateway:
     def __init__(self, host, port_expose, servicer, logger, proxy) -> None:
         with ImportExtensions(required=True):
             from fastapi import FastAPI, Body
-            from fastapi.encoders import jsonable_encoder
             from fastapi.responses import JSONResponse
             from fastapi.middleware.cors import CORSMiddleware
