+++ jina/peapods/headtail/head_pea.py
@@ -4,8 +4,8 @@ __license__ = "Apache-2.0"
 import argparse
 from typing import Dict, Union
 
-from .pea import BasePea
-from ..logging import JinaLogger
+from ..pea import BasePea
+from ...logging import JinaLogger
 
 
 class HeadPea(BasePea):
