+++ jina/peapods/pea.py
@@ -16,7 +16,8 @@ from jina.types.message import Message
 from .zmq import send_ctrl_message, Zmqlet, ZmqStreamlet
 from .. import __ready_msg__, __stop_msg__, Request
 from ..enums import PeaRoleType, SkipOnErrorType
-from ..excepts import RequestLoopEnd, NoExplicitMessage, ExecutorFailToLoad, MemoryOverHighWatermark, DriverError, PeaFailToStart, \
+from ..excepts import RequestLoopEnd, NoExplicitMessage, ExecutorFailToLoad, MemoryOverHighWatermark, DriverError, \
+    PeaFailToStart, \
     ChainedPodException
 from ..executors import BaseExecutor
 from ..helper import is_valid_local_config_source, typename
@@ -136,7 +137,8 @@ class BasePea(metaclass=PeaMeta):
         if 'role' in self.args and self.args.role == PeaRoleType.PARALLEL:
             self.name = f'{self.name}-{self.args.pea_id}'
         if 'host' in self.args and 'port_ctrl' in self.args and 'ctrl_with_ipc' in self.args:
-            self.ctrl_addr, self.ctrl_with_ipc = Zmqlet.get_ctrl_address(self.args.host, self.args.port_ctrl, self.args.ctrl_with_ipc)
+            self.ctrl_addr, self.ctrl_with_ipc = Zmqlet.get_ctrl_address(self.args.host, self.args.port_ctrl,
+                                                                         self.args.ctrl_with_ipc)
 
         if 'log_id' in self.args and 'log_config' in self.args:
             self.logger = JinaLogger(self.name,
